fix: move time&random to ecs
