+++ game_source/ECS/Time.cpp
@@ -1,4 +1,4 @@
-#include "Time.hpp"
+#include "include/Time.hpp"
 
 namespace Time {
     float deltaTime = 0; // Definition of the variable.
